Refactored the JaiaLog API into its own file
