--- src/plot/client/index.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-
-    <title>Logs - Jaia</title>
-    <link rel="icon" type="image/x-icon" href="/favicon.png">
-    <link rel="stylesheet" href="style.css">
-
-  </head>
-
-  <body id="body">
-
-    <div class="maximized">
-      <div class="vertical flexbox">
-
-        <div class="row" style="background-color: white;">
-          <div>
-            <image src="/favicon.png" style="display: inline-block;"></image>
-            <h2 style="display: inline-block; vertical-align: text-bottom; margin: 10pt">Logs</h2>
-          </div>
-        </div>
-
-        <div>
-          <label for="log" class="label padded">Log</label>
-          <select name="log" id="log" onchange="log_was_selected(event)">
-          </select>
-        </div>
-
-        <div>
-          <label for="path" class="label padded">Path</label>
-          <div id="path_div"></div>
-          <select name="path" id="path" onchange="path_was_selected(event)">
-          </select>
-        </div>
-      </div>
-
-      <div class="bottom_pane">
-        <div class="plotcontainer">
-          <div id="plot" class="plot"></div>
-        </div>
-
-        <div class="map" id="map"></div>
-      </div>
-
-    </div>
-
-  </body>
-
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-  <script src="https://cdn.plot.ly/plotly-2.12.1.min.js"></script>
-  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
-  integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A=="
-  crossorigin=""/>
-  <!-- Make sure you put this AFTER Leaflet's CSS -->
-  <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"
-    integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA=="
-    crossorigin=""></script>
-
-  <script src="script.js"></script>
-
-</html>
