--- src/plot/jaia_paths.json
@@ -1,27 +1,0 @@
-[
-    {
-        "path_regex": "jaiabot::bot_status;[0-9]+/jaiabot.protobuf.BotStatus/mission_state"
-    },
-    {
-        "units": "°C",
-        "path_regex": "jaiabot::bot_status;[0-9]+/jaiabot.protobuf.BotStatus/temperature"
-    },
-    {
-        "units": "m",
-        "path_regex": "jaiabot::bot_status;[0-9]+/jaiabot.protobuf.BotStatus/depth"
-    },
-    {
-        "units": "°",
-        "path_regex": "jaiabot::bot_status;[0-9]+/jaiabot.protobuf.BotStatus/location/lat",
-        "invalid_values": [0.0]
-    },
-    {
-        "units": "°",
-        "path_regex": "jaiabot::bot_status;[0-9]+/jaiabot.protobuf.BotStatus/location/lon",
-        "invalid_values": [0.0]
-    },
-    {
-        "units": "μsec",
-        "path_regex": ".+ArduinoCommand/motor"
-    }
-]
